requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
drf-spectacular>=0.27,<0.28
pymongo>=4.6,<4.7
django-environ>=0.11,<0.12
django-redis>=5.4,<5.5
redis>=5.0,<6
psycopg[binary]>=3.1.10,<3.2
uvicorn>=0.23,<0.24
gunicorn>=21.2,<22
whitenoise>=6.6,<6.7
